# Faster type and symbol resolution