--- a/cpg-core/src/main/kotlin/de/fraunhofer/aisec/cpg/passes/SymbolResolver.kt
+++ b/cpg-core/src/main/kotlin/de/fraunhofer/aisec/cpg/passes/SymbolResolver.kt
@@ -421,6 +421,16 @@ open class SymbolResolver(ctx: TranslationContext) : ComponentPass(ctx) {
         val callee = call.callee
         val language = call.language
 
+        // If the base type is unknown, we cannot resolve the call
+        if (callee is MemberExpression && callee.base.type is UnknownType) {
+            Util.warnWithFileLocation(
+                call,
+                log,
+                "Cannot resolve call to ${callee.name} because the base type is unknown",
+            )
+            return
+        }
+
         // Handle a possible overloaded operator->
         resolveOverloadedArrowOperator(callee)