--- a/cpg-core/src/test/kotlin/de/fraunhofer/aisec/cpg/passes/scopes/ScopeManagerTest.kt
+++ b/cpg-core/src/test/kotlin/de/fraunhofer/aisec/cpg/passes/scopes/ScopeManagerTest.kt
@@ -82,7 +82,7 @@ internal class ScopeManagerTest : BaseTest() {
             assertNotNull(scopeA)
 
             // should also be able to look up via the FQN
-            assertEquals(scopeA, final.lookupScope("A"))
+            assertEquals(scopeA, final.lookupScope(parseName("A")))
 
             // and it should contain both functions from the different file in the same namespace
             assertContains(scopeA.symbols["func1"] ?: listOf(), func1)
@@ -112,28 +112,29 @@ internal class ScopeManagerTest : BaseTest() {
         val frontend =
             TestLanguageFrontend("::", TestLanguage(), TranslationContext(config, s, TypeManager()))
         s.resetToGlobal(frontend.newTranslationUnitDeclaration("file.cpp", null))
+        with(frontend) {
+            assertNull(s.currentNamespace)
 
-        assertNull(s.currentNamespace)
+            val namespaceA = frontend.newNamespaceDeclaration("A", null)
+            s.enterScope(namespaceA)
 
-        val namespaceA = frontend.newNamespaceDeclaration("A", null)
-        s.enterScope(namespaceA)
+            assertEquals("A", s.currentNamespace.toString())
 
-        assertEquals("A", s.currentNamespace.toString())
+            // nested namespace A::B
+            val namespaceB = frontend.newNamespaceDeclaration("B", null)
+            s.enterScope(namespaceB)
 
-        // nested namespace A::B
-        val namespaceB = frontend.newNamespaceDeclaration("B", null)
-        s.enterScope(namespaceB)
+            assertEquals("A::B", s.currentNamespace.toString())
 
-        assertEquals("A::B", s.currentNamespace.toString())
+            val func = frontend.newFunctionDeclaration("func")
+            s.addDeclaration(func)
 
-        val func = frontend.newFunctionDeclaration("func")
-        s.addDeclaration(func)
+            s.leaveScope(namespaceB)
+            s.addDeclaration(namespaceB)
+            s.leaveScope(namespaceA)
 
-        s.leaveScope(namespaceB)
-        s.addDeclaration(namespaceB)
-        s.leaveScope(namespaceA)
-
-        val scope = s.lookupScope("A::B")
-        assertNotNull(scope)
+            val scope = s.lookupScope(parseName("A::B"))
+            assertNotNull(scope)
+        }
     }
 }