This PR speeds up type and symbol resolution across the whole code-base

* Optimized `extractScope` by using the `fqnScopeMap` instead of filtering ALL scopes.
* Removed unnecessary nested loop over all nodes/types again in `TypeManager`
* Faster type/scope merging in the `TranslationManager` the using parallel parsing (8ms instead of 5s)
* Do not try to resolve member expressions of an unknown base type. This will fail and will only result in a lot of spam and in the attempts to resolve the "UNKNOWN" scope (which does not exist)

This boosted my test case (parsing python type shed stdlib) from ~70s to ~~45s~~ 32s